[F] Always ensure nginx executable status
The ensure-nginx-executable process makes it so that Nginx can be bound
to 80 and 443 when run by a normal user. It really only needs to be run
after Nginx has been updated by homebrew, but for now, this should cover
it. The script doesn't ask for sudo or do anything unless it determines
it needs to run (thanks @scryptmouse)